eve of WCUS

+++ pzprojectgrid/pzprojectgrid.php
@@ -13,14 +13,8 @@
  * @package           pzprojectgrid
  */
 
-/**
- * Registers the block using the metadata loaded from the `block.json` file.
- * Behind the scenes, it registers also all assets so they can be enqueued
- * through the block editor in the corresponding context.
- *
- * @see https://developer.wordpress.org/reference/functions/register_block_type/
- */
-
+//normally this line will be commented out
+defined('DEMO') || define('DEMO', true);
 
 
 function create_block_pzprojectgrid_block_init() {
@@ -44,7 +38,7 @@ function pz_project_grid_block($attributes) {
 
 	ob_start();
 	?>
-	<div class="pz-projectgrid-div"><pre><?php echo wp_json_encode($attributes); ?></pre></div>
+	<div class="pz-projectgrid-div" ><pre><?php echo wp_json_encode($attributes); ?></pre></div>
 	
 	<?php
 	return ob_get_clean();
@@ -53,7 +47,10 @@ function pz_project_grid_block($attributes) {
 function do_project_delete() {
 	global $wpdb;
 
-	$wpdb->delete( $wpdb->prefix . 'pz_project', array( 'id' => $_POST['id'] ) );
+	if( ! defined('DEMO')) {
+		$wpdb->delete( $wpdb->prefix . 'pz_project', array( 'id' => $_POST['id'] ) );
+	}
+	
 	wp_redirect( $_POST['postDeleteURL']);
 	exit;
